+++ src/OpenTelemetry/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Prevent accessing activity Id before sampler runs in case of legacy
+  activities.
+  ([2659](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2659))
+
 * Added `ReadOnlyTagCollection` and expose `Tags` on `MetricPoint` instead of
   `Keys`+`Values`
   ([#2642](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2642))
@@ -20,6 +24,15 @@
   enumerating `BucketCounts` and `ExplicitBounds`. Removed `GetBucketCounts` and
   `GetExplicitBounds` methods from `MetricPoint`.
   ([#2664](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2664))
+
+* Refactored temporality setting to align with the latest spec.
+  ([#2666](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2666))
+
+* Removed the public properties `LongValue`, `DoubleValue`, `StartTime`, and
+  `EndTime` in favor of their counterpart public methods `GetCounterSumLong`,
+  `GetCounterSumDouble`, `GetGaugeLastValueLong`, `GetGaugeLastValueDouble`,
+  `GetStartTime`, and `GetEndTime`.
+  ([#2667](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2667))
 
 ## 1.2.0-beta2